driver: Evaluate and maintain Linux' __boot_cpu_mode on ARM
We need to clear the BOOT_CPU_MODE_MISMATCH flag from __boot_cpu_mode
when disabling Jailhouse. It will be set by Linux when online a CPU
while Jailhouse is running, because we then let the guest start in SVC
instead of HYP mode. If we leave the flag set, the next physical CPU
onlining will not properly install the hypervisor vectors, and we will
crash when trying to enable Jailhouse again.

We no rely on a tiny upstream Linux patch to export __boot_cpu_mode to
GPL modules. As it now becomes available for us, we can also use it (via
is_hyp_mode_available) to check for the availability of the hypervisor
stub Linux should have installed instead of simply crashing the system
when it is missing.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>