arm: Align description arch_paging_flush_cpu_caches with actual logic
Despite claiming to perform also a dcache invalidation,
arch_paging_flush_cpu_caches was only doing a clean so far, probably due
to a mistake in the definition of DCCIMVAC (it was defined as DCCMVAC).

However, there is no need to invalidate the caches here. Therefore align
the sysreg definition and comments to what the code already did.

CC: Marc Zyngier <marc.zyngier@arm.com>
CC: Mark Rutland <mark.rutland@arm.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>