arm: Rename ESR to HSR
In the attempt to reduce the mapping effort on AArch64, HSR was called
ESR_EL2, like on AArch64. Also all related constants where prefixed with
ESR. This is only confusing when looking at the 32-bit spec. As arm64
will come with its own constants anyway, clean up the misleading names
on arm.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>