arm: Prepare for trapping virtual memory control registers
This adds all trap handling required once HCR.TVM is set. Then a number
of system control register writes trap to HYP mode, and we need to
execute them on behalf of the guest.

The only one we are actually interested in is SCTLR, so this trap is
already prepare for additional logic. The others are simply handled via
a macro that encapsulates the match as well as the register write.
TTBR0/1 may either show up as 32 or 64-bit write, depending on LPAE
being off or on.

CC: Marc Zyngier <marc.zyngier@arm.com>
CC: Mark Rutland <mark.rutland@arm.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>